CRM-17860 - CiviTester - Split into separate class files
