+++ Civi/Test/CiviEnvBuilder/StepInterface.php
@@ -1,0 +1,11 @@
+<?php
+namespace Civi\Test\CiviEnvBuilder;
+
+interface StepInterface {
+  public function getSig();
+
+  public function isValid();
+
+  public function run($ctx);
+
+}
